Silenced a signed/unsigned comparison warning
svn-id: r32355

+++ engines/made/screen.cpp
@@ -819,7 +819,7 @@ int16 Screen::addToSpriteList(int16 index, int16 xofs, int16 yofs) {
 }
 
 SpriteListItem Screen::getFromSpriteList(int16 index) {
-	if (index > _spriteList.size()) {
+	if (((uint) index) > _spriteList.size()) {
 		SpriteListItem emptyItem;
 		emptyItem.index = 0;
 		return emptyItem;
